+++ changelog.md
@@ -35,6 +35,7 @@ Codion Change Log
 - FilterListModel and FilterListSort added.
 - FilterComboBoxModel.Builder.asyncRefresh() added.
 - FilterListModel, FilterTableModel.Builder.visible() added.
+- ProgressWorker.Task and ProgressTask added, related changes.
 ### is.codion.swing.common.ui
 - LookAndFeelSelectionDialogBuilder.createControl() overloaded without parameter.
 - BackgroundColorValidIndicator bug fixed, did not refresh state when colors changed during updateUI().